requirements.txt
tqdm
wheel
pycocotools
scikit-learn
requests
inflect
fiftyone
fairscale
sentencepiece
modelscope
accelerate
diffusers
click
open_clip_torch
clip
promptcap
salesforce-lavis
transformers==4.30.2
ninja